cleanup

+++ src/coding/coding.h
@@ -167,7 +167,7 @@ void decode_msadpcm_stereo(VGMSTREAM* vgmstream, sample_t* outbuf, int32_t first
 void decode_msadpcm_mono(VGMSTREAM* vgmstream, sample_t* outbuf, int channelspacing, int32_t first_sample, int32_t samples_to_do, int channel, int config);
 void decode_msadpcm_ck(VGMSTREAM* vgmstream, sample_t* outbuf, int channelspacing, int32_t first_sample, int32_t samples_to_do, int channel);
 long msadpcm_bytes_to_samples(long bytes, int block_size, int channels);
-int msadpcm_check_coefs(STREAMFILE* sf, off_t offset);
+int msadpcm_check_coefs(STREAMFILE* sf, uint32_t offset);
 
 
 /* yamaha_decoder */
